move TemplateType to index
