--- libs/remix-ui/workspace/src/lib/utils/types.ts
@@ -1,8 +1,0 @@
-export type TemplateType = {
-    type: 'git' | 'plugin'
-    url?: string
-    branch?: string
-    name?: string
-    endpoint?: string
-    params?: any[]
-  }
